Implemented additional objects: parallelopiped [this has a data structure in ppcom.f] coordinate-aligned cube coordinate-aligned cylinder and tested to show that the cijroutin works properly for them. The solutions look qualitatively sensible. Added some test data files geomcube.dat geomcyl.dat Changed the color gradient in accis/slicing.f

+++ average.f
@@ -8,7 +8,7 @@
          do j=1,iuds(2)
             do i=1,iuds(1)
                qave(i,j,k)=((istepave-1)*qave(i,j,k)+q(i,j,k))
-     $              /int(istepave)
+     $              /float(istepave)
             enddo
          enddo
       enddo
